fix: mv playwright config from ts to mjs
